feature: card take options
