+++ lib/avo/dashboards/base_divider.rb
@@ -3,12 +3,14 @@ module Avo
     class BaseDivider
       attr_reader :label
       attr_reader :invisible
+      attr_reader :index
 
       class_attribute :id
 
-      def initialize(label: nil, invisible: false)
+      def initialize(label: nil, invisible: false, index: nil)
         @label = label
         @invisible = invisible
+        @index = index
       end
 
       def is_divider?
