+++ lib/avo/dashboards/dashboard_card.rb
@@ -1,0 +1,6 @@
+module Avo
+  module Dashboards
+    class DashboardCard < Avo::BaseCard
+    end
+  end
+end
