+++ spec/dummy/app/views/avo/cards/_custom_card.html.erb
@@ -1,7 +1,13 @@
 <div class="flex-1 flex p-4">
   <div class="place-self-end">
-    Dashboard ID: <%= @dashboard.id %>
+    Dashboard ID: <%= @dashboard.inspect %>
     <br>
     Current user ID: <%= current_user.id %>
+    <br>
+    Params: <pre><%= params.inspect %></pre>
+    <br>
+    Options: <%= @card.options %>
+    <br>
+    Options block: <%= @card.options[:block].call params %>
   </div>
 </div>
